Improve DRNSF.exe search mechanism
Prior to this change, CE merely attempts to invoke a program by the name
of "drnsf" within the CE EXE directory. This works as long as the drnsf
executable (or an appropriate symlink) is in the directory, but does not
work for other directory layouts.

This change modifies the search routine to check for any file named
"drnsf" (case-insensitive) plus any optional file extension (this allows
again for exe, cmd, bat, etc, on windows) in the following locations:

  * The main directory
  * The parent directory
  * Any subdirectory named "drnsf" (case-insensitive) of one of the
    above directories.

For example, given the CE filename below:

  Z:\ModTools\CB\CrashEdit.exe

The following files are all considered:

  * Z:\ModTools\CB\DRNSF.exe
  * Z:\ModTools\CB\DRNSF\DRNSF.exe
  * Z:\ModTools\DRNSF.exe
  * Z:\ModTools\DRNSF\DRNSF.exe

All file extensions are considered (.exe, .bat, .cmd, .slewp, none). The
first file found may or may not be executable.

Linux symbolic links appear to function OK on Mono.

This change also fixes an error message which was intended to display
when DRNSF could not be located. It did not display prior to this
change, but does now.