Completely rewrite C2 animation system
Model entry code has been re-written to be much more comprehensible and better overall. The vertex decompression has also been reworked and should even be slightly faster. Polygon resolving has been moved to the model entry load, previously it resolved the polygon data every time it wanted to render the animation, even though it's independent from the current frame, which lead to (luckily not so serious) performance issues. The code was very hacky and the variables weren't well-named, so it just looked like a giant mess of horrible code that somehow worked. The new code isn't as good as I'd like it to me, but it's still much better, and is way faster. However, it's incomplete for now, as it doesn't do color resolving yet. Therefore, it will currently render Crash 2 animations in wireframe.