Update CrashEdit program form styles
This changes the style of most controls (buttons, spinboxes, text, etc.) to be based more on system instead of the framework. Generally this should make things look a lot nicer or at least themed to the OS it's running on.

This is also how forms programs are set up by default in VS 2005 and later.