Fixed issue loading up the Warp room Zone 10_2Z.
The Zone would fail and break the Zone Viewer because the Scenery to it (2__2W) was broken due to this line. This from visual tests seems to work I think.

However why is the monitor on the top of the area not animated like it does in game? Maybe these are the animated textures that are failing.